+++ tests/kitchen/kitchenTest1/Default/Get-LsbRelease.tests.ps1
@@ -1,5 +1,0 @@
-Describe 'Get-LsbRelease function' -skip:$IsWindows {
-    it 'should not throw (if the binary is installed)' {
-     {Get-LsbRelease} | Should -not -Throw
-    }
-}